requirements.txt
Django==1.8.3
daphne==1.0.1
mysqlclient==1.3.9
channels==1.0.2
djangorestframework==3.1.3
asgi-redis==1.0.0